requirements.txt
pandas
requests
google-cloud-bigquery
pyarrow
fastapi
uvicorn